Fixed API connection

+++ src/main/java/com/example/cryptotracker/model/Cryptocurrency.java
@@ -5,11 +5,12 @@ import java.math.BigInteger;
 public class Cryptocurrency {
     private String id;
     private String market_cap_rank;
-    private int current_price;
+    private double current_price;
     private BigInteger market_cap;
     private double market_cap_change_percentage_24h;
 
-    public Cryptocurrency(String id, String market_cap_rank, int current_price, BigInteger market_cap, double market_cap_change_percentage_24h) {
+    // Costruttore
+    public Cryptocurrency(String id, String market_cap_rank, double current_price, BigInteger market_cap, double market_cap_change_percentage_24h) {
         this.id = id;
         this.market_cap_rank = market_cap_rank;
         this.current_price = current_price;
@@ -17,7 +18,7 @@ public class Cryptocurrency {
         this.market_cap_change_percentage_24h = market_cap_change_percentage_24h;
     }
 
-
+    // Getter e Setter per id
     public String getId() {
         return id;
     }
@@ -26,6 +27,7 @@ public class Cryptocurrency {
         this.id = id;
     }
 
+    // Getter e Setter per market_cap_rank
     public String getMarket_cap_rank() {
         return market_cap_rank;
     }
@@ -34,14 +36,16 @@ public class Cryptocurrency {
         this.market_cap_rank = market_cap_rank;
     }
 
-    public int getCurrent_price() {
+    // Getter e Setter per current_price
+    public double getCurrent_price() {
         return current_price;
     }
 
-    public void setCurrent_price(int current_price) {
+    public void setCurrent_price(double current_price) {
         this.current_price = current_price;
     }
 
+    // Getter e Setter per market_cap
     public BigInteger getMarket_cap() {
         return market_cap;
     }
@@ -50,6 +54,7 @@ public class Cryptocurrency {
         this.market_cap = market_cap;
     }
 
+    // Getter e Setter per market_cap_change_percentage_24h
     public double getMarket_cap_change_percentage_24h() {
         return market_cap_change_percentage_24h;
     }
